+++ src/extension.ts
@@ -37,7 +37,7 @@ export function activate(context: ExtensionContext) {
 
     // Create the language client and start the client.
     client = new LanguageClient(
-        'ageOfEmpires2AI',
+        'aoe2ai',
         'Age Of Empires II AI',
         serverOptions,
         clientOptions